added chunk and block classes

+++ src/Vector3Float.java
@@ -1,17 +1,11 @@
 
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-//package basic3d;
-
 /**
  *
  * @author sarmi
  */
 public class Vector3Float {
     public float x, y, z;
-    public Vector3Float(int x, int y, int z){
+    public Vector3Float(float x, float y, float z){
         this.x=x;
         this.y=y;
         this.z=z;
